+++ client/src/containers/PlannedRecipe.js
@@ -1,11 +1,11 @@
 import PropTypes from "prop-types";
 import React from "react";
-import Dispatcher from "../data/dispatcher";
+import Dispatcher from "data/dispatcher";
 import LibraryStore from "features/RecipeLibrary/data/LibraryStore";
-import TaskActions from "../data/TaskActions";
-import TaskStore from "../data/TaskStore";
-import useFluxStore from "../data/useFluxStore";
-import LoadObject from "../util/LoadObject";
+import TaskActions from "features/Planner/data/TaskActions";
+import TaskStore from "features/Planner/data/TaskStore";
+import useFluxStore from "data/useFluxStore";
+import LoadObject from "util/LoadObject";
 import LoadingIndicator from "../views/common/LoadingIndicator";
 import RecipeDetail from "../views/cook/RecipeDetail";
