+++ client/src/containers/RecipeEdit.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { withRouter } from "react-router-dom";
-import useIngredientLO from "../data/useIngredientLO";
+import useIngredientLO from "data/useIngredientLO";
 import RecipeEdit from "../views/cook/RecipeEdit";
 
 export default withRouter(({match}) => {
