+++ client/src/containers/Shop.js
@@ -1,16 +1,16 @@
 import React from "react";
 import LibraryStore from "features/RecipeLibrary/data/LibraryStore";
-import ShoppingStore from "../data/ShoppingStore";
+import ShoppingStore from "data/ShoppingStore";
 import {
     isParent,
     isQuestionable,
     isSection,
-} from "../data/tasks";
-import TaskStatus from "../data/TaskStatus";
-import TaskStore from "../data/TaskStore";
-import useFluxStore from "../data/useFluxStore";
-import groupBy from "../util/groupBy";
-import ShopList from "../views/shop/ShopList";
+} from "features/Planner/data/tasks";
+import TaskStatus from "features/Planner/data/TaskStatus";
+import TaskStore from "features/Planner/data/TaskStore";
+import useFluxStore from "data/useFluxStore";
+import groupBy from "util/groupBy";
+import ShopList from "views/shop/ShopList";
 
 const gatherLeaves = item => {
     if (!isParent(item)) {
