+++ client/src/data/PantryItemActions.js
@@ -1,5 +1,5 @@
 import PropTypes from "prop-types";
-import typedAction from "../util/typedAction";
+import typedAction from "util/typedAction";
 
 const PantryItemActions = {
     LOAD_PANTRYITEMS: "LOAD_PANTRYITEMS",
