+++ client/src/data/UserActions.js
@@ -1,5 +1,5 @@
 import PropTypes from "prop-types";
-import typedAction from "../util/typedAction";
+import typedAction from "util/typedAction";
 
 const UserActions = {
     RESTORE_PREFERENCES: "user/restore-preferences",
