+++ client/src/data/ValidatingDispatcher.js
@@ -1,7 +1,7 @@
 import { Dispatcher } from "flux";
 import invariant from "invariant";
 import PropTypes from "prop-types";
-import { CONTAINER_KEY } from "../util/typedAction";
+import { CONTAINER_KEY } from "util/typedAction";
 
 const checkPayload = payload => {
     invariant(payload.type != null, "Actions must have a 'type' key");
