+++ client/src/data/snackBarStore.js
@@ -2,14 +2,14 @@ import { Button } from "@material-ui/core";
 import { ReduceStore } from "flux/utils";
 import PropTypes from "prop-types";
 import React from "react";
-import typedStore from "../util/typedStore";
+import typedStore from "util/typedStore";
 import dispatcher from "./dispatcher";
 import LibraryStore from "features/RecipeLibrary/data/LibraryStore";
 import PantryItemActions from "./PantryItemActions";
 import RecipeActions from "./RecipeActions";
-import TaskActions from "./TaskActions";
-import { willStatusDelete } from "./TaskStatus";
-import TaskStore from "./TaskStore";
+import TaskActions from "features/Planner/data/TaskActions";
+import { willStatusDelete } from "features/Planner/data/TaskStatus";
+import TaskStore from "features/Planner/data/TaskStore";
 import UiActions from "./UiActions";
 
 const enqueue = (state, item) => {
