+++ client/src/features/Planner/data/TaskStore.js
@@ -2,34 +2,34 @@ import dotProp from "dot-prop-immutable";
 import { ReduceStore } from "flux/utils";
 import invariant from "invariant";
 import PropTypes from "prop-types";
-import { removeAtIndex } from "../util/arrayAsSet";
-import ClientId, { clientOrDatabaseIdType } from "../util/ClientId";
-import { bucketComparator } from "../util/comparators";
-import inTheFuture from "../util/inTheFuture";
-import LoadObject from "../util/LoadObject";
-import LoadObjectState from "../util/LoadObjectState";
+import { removeAtIndex } from "util/arrayAsSet";
+import ClientId, { clientOrDatabaseIdType } from "util/ClientId";
+import { bucketComparator } from "util/comparators";
+import inTheFuture from "util/inTheFuture";
+import LoadObject from "util/LoadObject";
+import LoadObjectState from "util/LoadObjectState";
 import {
     loadObjectOf,
     loadObjectStateOf,
-} from "../util/loadObjectTypes";
-import socket from "../util/socket";
+} from "util/loadObjectTypes";
+import socket from "util/socket";
 import {
     formatLocalDate,
     parseLocalDate,
-} from "../util/time";
-import typedStore from "../util/typedStore";
-import AccessLevel from "./AccessLevel";
-import Dispatcher from "./dispatcher";
-import PantryItemActions from "./PantryItemActions";
-import PreferencesStore from "./PreferencesStore";
-import ShoppingActions from "./ShoppingActions";
-import TaskActions from "./TaskActions";
-import TaskApi from "./TaskApi";
+} from "util/time";
+import typedStore from "util/typedStore";
+import AccessLevel from "data/AccessLevel";
+import Dispatcher from "data/dispatcher";
+import PantryItemActions from "data/PantryItemActions";
+import PreferencesStore from "data/PreferencesStore";
+import ShoppingActions from "data/ShoppingActions";
+import TaskActions from "features/Planner/data/TaskActions";
+import TaskApi from "features/Planner/data/TaskApi";
 import {
     isExpanded,
     isParent,
-} from "./tasks";
-import TaskStatus, { willStatusDelete } from "./TaskStatus";
+} from "features/Planner/data/tasks";
+import TaskStatus, { willStatusDelete } from "features/Planner/data/TaskStatus";
 
 /*
  * This store is way too muddled. But leaving it that way for the moment, to
