+++ client/src/providers/Profile.js
@@ -11,7 +11,7 @@ import {
     LOCAL_STORAGE_ACCESS_TOKEN,
 } from "../constants";
 import GTag from "../GTag";
-import LoadObject from "../util/LoadObject";
+import LoadObject from "util/LoadObject";
 
 // global side effect to ensure cookies are passed
 BaseAxios.defaults.withCredentials = true;
