+++ client/src/views/Landing.js
@@ -2,7 +2,7 @@ import PropTypes from "prop-types";
 import React from "react";
 import { Redirect } from "react-router-dom";
 import { SESSION_STORAGE_POST_LOGIN } from "../constants/index";
-import { getJsonItem } from "../util/storage";
+import { getJsonItem } from "util/storage";
 
 const Landing = ({authenticated}) => {
     let dest = getJsonItem(SESSION_STORAGE_POST_LOGIN, sessionStorage);
