+++ client/src/views/cook/SubrecipeItem.js
@@ -7,8 +7,8 @@ import makeStyles from "@material-ui/core/styles/makeStyles";
 import PropTypes from "prop-types";
 import React from "react";
 import { Recipe } from "../../data/RecipeTypes";
-import { formatDuration } from "../../util/time";
-import CollapseIconButton from "../plan/CollapseIconButton";
+import { formatDuration } from "util/time";
+import CollapseIconButton from "global/components/CollapseIconButton";
 import IngredientDirectionsRow from "./IngredientDirectionsRow";
 
 const useStyles = makeStyles({
